Refactor handlers and add cloud run container
