+++ functions/.dockerignore
@@ -1,0 +1,17 @@
+node_modules
+npm-debug.log
+.git
+.gitignore
+README.md
+.env
+.env.*
+test
+coverage
+.nyc_output
+*.test.js
+jest.config.js
+.eslintrc*
+.prettierrc*
+tsconfig.json
+firebase.json
+.firebaserc
