+++ functions/Dockerfile
@@ -1,0 +1,15 @@
+FROM node:20-slim
+
+WORKDIR /app
+
+# Copy package files
+COPY package*.json ./
+
+# Install dependencies
+RUN npm ci --only=production
+
+# Copy application files
+COPY . ./
+
+# Start the application
+CMD ["node", "webapp.js"]
