+++ functions/package.json
@@ -7,7 +7,7 @@
     "typecheck": "tsc --noEmit",
     "serve": "firebase serve --only functions",
     "shell": "firebase functions:shell",
-    "start": "npm run shell",
+    "start": "node webapp.js",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log",
     "test": "jest",
@@ -16,6 +16,7 @@
   "dependencies": {
     "@google-cloud/logging": "^11.0.0",
     "@valkey/valkey-glide": "^2.0.1",
+    "fastify": "^4.28.1",
     "firebase-admin": "^12.1.0",
     "firebase-functions": "^5.0.1"
   },
